requirements.txt
Django==1.5.1
argparse==1.2.1
coverage==3.6
distribute==0.6.35
httplib2==0.8
mock==1.0.1
stevedore==0.8
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==3.7
wsgiref==0.1.2
